src/rpi/requirements.txt
opencv-python~=4.10.0.82
numpy~=1.26.4
pillow~=10.2.0
pynput~=1.7.7
python-vlc~=3.0.20123
pyserial~=3.5
scipy~=1.11.4
matplotlib~=3.8.3
pandas~=1.5.3
tflite-runtime~=2.14.0
luma.oled~=3.13.0
requests
python-dotenv
pygame
SpeechRecognition
mutagen
pyaudio